% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_Hive.R
\name{as.data.frame.tbl_Hive}
\alias{as.data.frame.tbl_Hive}
\title{Pull rows from a Hive table into a local data.frame}
\usage{
\method{as.data.frame}{tbl_Hive}(x, n = 100, ...)
}
\arguments{
\item{x}{A tbl_Hive object}

\item{n}{Number of rows to collect}
}
\description{
\code{as.data.frame.tbl_Hive} collects rows from a 
Hive table and coerces it to a data.frame
}
\details{
The reason this function was created was to fix a problem
when printing Hive tables with a large number of columns
}

