% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_Hive.R
\name{collect}
\alias{collect}
\alias{collect.tbl_Hive}
\title{Pull a Hive table to a local data frame}
\usage{
\method{collect}{tbl_Hive}(x, n, batch = 1e+05, quiet = TRUE, ...)
}
\arguments{
\item{x}{A \code{tbl_Hive} object}

\item{n}{The number of rows to pull down}

\item{batch}{Size of batches in which to fetch the data (passed to 
\code{\link[orpheus]{hive_query}})}

\item{quiet}{Whether to print progress updates as the data is being fetched}

\item{...}{Additional arguments passed to \code{\link[orpheus]{hive_query}}}
}
\value{
A tibble of the result
}
\description{
\code{collect} forces computation of a query and pulls the result 
down into a local data frame.
}
\details{
By default, collect will pull down the 100,000 rows of the result table.
If the result table pulls back exactly 100,000 rows, 
a warning message will be printed.

To pull down all of the rows, \code{n = Inf} can be specified. 
When using \code{collect}, make sure to keep in mind how large the data 
set you're pulling down is, in regard to both the number of rows as well 
as the number of columns.

\code{collect} works by calling \code{hive_query}, so if necessary, you can specify 
the batch argument like you would in hive_query to avoid out-of-memory
errors, and you can specify the \code{quiet} argument for whether to print 
update messages as the data is being pulled. In contrast to \code{hive_query}, 
for \code{collect} the \code{quiet} parameter is TRUE by default. 
You can set \code{quiet = FALSE} if you want messages to be printed.

To the extent that you can leave the data in Hive, it is best to do so. 
\code{collect} should only be called once you have a data set that has been 
filtered, aggregated, and narrowed down to the columns you need, 
such as a modeling data set.
}

