% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_Hive.R
\name{compute}
\alias{compute}
\alias{compute.tbl_Hive}
\title{Compute a Hive table}
\usage{
\method{compute}{tbl_Hive}(x, name = dplyr:::random_table_name(),
  temporary = TRUE, ..., overwrite = FALSE)
}
\arguments{
\item{x}{A \code{tbl_Hive} object}

\item{name}{The name of the Hive table to create}

\item{temporary}{Whether to create a temporary or permanent table}

\item{overwrite}{Whether to overwrite an existing table with the
same name if one exists in the Hive database}
}
\value{
A \code{tbl_Hive} object pointing to the resulting table
}
\description{
\code{compute} forces computation of a Hive table, saving the result in
a Hive database.
}
\details{
By default, a temporary table in the Hive database will 
be created, unless you specify that it should be a permanent table.

As a general rule, it is good practice to call compute after every few steps.
Doing so helps make troubleshooting easier, as you don't need to wade 
through deeply nested statements to figure out what is going on, making 
it simpler to track down the point at something is going awry.
}

