% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_Hive.R
\name{copy_to}
\alias{copy_to}
\alias{copy_to.src_Hive}
\title{Copy a local data frame to Hive}
\usage{
\method{copy_to}{src_Hive}(src, df, name = lazyeval::expr_text(df),
  temporary = TRUE, indexes = NULL, overwrite = FALSE)
}
\arguments{
\item{src}{A \code{src_Hive} connection object}

\item{df}{The local data frame to copy}

\item{name}{What to name the table in Hive}

\item{temporary}{Whether to create a temporary or permanent table}

\item{overwrite}{Whether to overwrite an existing table with the
same name if one exists in the Hive database}
}
\value{
A \code{tbl_Hive} object reference to the Hive table created
}
\description{
\code{copy_to} copies a local data frame to 
a table in the active Hive database.
}
\details{
By default, a temporary table in the Hive database will be created, 
unless you specify that it should be a permanent table.
}

