% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{gather_hive}
\alias{gather_hive}
\title{Gather columns into key-value pairs}
\usage{
gather_hive(hive_tbl, key, value, ..., temporary = TRUE,
  name = dplyr:::random_table_name())
}
\arguments{
\item{hive_tbl}{A tbl_Hive object}

\item{key}{Bare unquoted name of the key column to create}

\item{value}{Bare unquoted name of the value column to create}

\item{...}{Specification of the columns to gather, using dplyr \code{\link[dplyr]{select}} syntax}

\item{temporary}{Whether the resulting table should be a temporary table or not}

\item{name}{Name of the resulting Hive table}
}
\description{
\code{gather_hive} operates on a Hive table, taking multiple columns and gathering them into key-value pairs
}

