% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_Hive.R
\name{hive_to_hdfs_txt}
\alias{hive_to_hdfs_txt}
\title{Export a Hive table to a txt file in HDFS}
\usage{
hive_to_hdfs_txt(hive_tbl, hdfs_path, delim = "\\t")
}
\arguments{
\item{hive_tbl}{A \code{tbl_Hive} object}

\item{hdfs_path}{HDFS path for the resulting file}

\item{delim}{Delimiter for the exported text file}
}
\description{
\code{hive_to_hdfs_txt} will export a Hive table to a delimited text file
in HDFS (tab-delimited by default)
}

