% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{spread_hive}
\alias{spread_hive}
\title{Spread a key-value pair across multiple columns}
\usage{
spread_hive(hive_tbl, key, value, temporary = TRUE,
  name = dplyr:::random_table_name())
}
\arguments{
\item{hive_tbl}{A tbl_Hive object}

\item{key}{Bare unquoted name of the key column, whose values will be column names}

\item{value}{Bare unquoted name of the value column, whose values will populate the cells}

\item{temporary}{Whether the resulting table should be a temporary table or not}

\item{name}{Name of the resulting Hive table}
}
\description{
\code{spread_hive} operates on a Hive table, taking a key and value column and spreading them to multiple columns
}

