% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_Hive.R
\name{src_Hive}
\alias{src_Hive}
\title{Create Hive connection to research cluster}
\usage{
src_Hive(db, ...)
}
\arguments{
\item{db}{Name of the Hive database}

\item{...}{Additional arguments passed to \code{\link[orpheus]{hive_connect}}}
}
\description{
\code{src_Hive} creates a connection to the research lake.
}
\details{
This is the connection that is used by dplyr to use Hive
as a backend.
}

