% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_Hive.R
\name{tbl}
\alias{tbl}
\alias{tbl.src_Hive}
\title{Establish a connection to a Hive table}
\usage{
\method{tbl}{src_Hive}(src, from, ...)
}
\arguments{
\item{src}{A src_Hive connection object}

\item{from}{Information to create the table connection (either a string of
the table name or a query wrapped in \code{sql})}
}
\description{
\code{tbl} is a method for the generic \code{\link[dplyr]{tbl}}
}
\details{
To create a connection to a Hive table in a database other than the one 
connected to in \code{src_Hive}, simply add the database name when 
specifying the table name.

In addition to specifying table names in a \code{tbl} statement, you can use 
arbitrary SQL itself to refer to a table. In place of the table name 
in the \code{tbl} statement, simply wrap the desired SQL in 
a \code{sql} call.
}

